+++ src/noahtheduke/splint.clj
@@ -28,6 +28,7 @@
    noahtheduke.splint.rules.lint.duplicate-case-test
    noahtheduke.splint.rules.lint.duplicate-field-name
    noahtheduke.splint.rules.lint.fn-wrapper
+   noahtheduke.splint.rules.lint.identical-branches
    noahtheduke.splint.rules.lint.if-else-nil
    noahtheduke.splint.rules.lint.if-let-else-nil
    noahtheduke.splint.rules.lint.if-nil-else